backend/worker/requirements.txt
celery
python-dotenv
psycopg2-binary
redis
openai
boto3
requests
mutagen
